requirements.txt
apify-client
openai
gspread
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
flask
tavily-python
